Final changes for Testimonial section
