+++ config/content/Testimonials.js
@@ -1,0 +1,21 @@
+const testimonials = [
+  {
+    quote:
+      "Interesting ! Is there a way to display public repositories from organizations made by the user ?",
+    name: "Yannick",
+    title: "Full-Stack Developer",
+  },
+  {
+    quote:
+      "The idea is very cool :) Would love to be able to export the result as a PDF I can share",
+    name: "Matt El Mouktafi",
+    title: "Founder @Achiev",
+  },
+  {
+    quote: "Looks nice. Good luck to you today",
+    name: "Garen Orchyan",
+    title: "Building Asyncinterview",
+  },
+];
+
+export default testimonials;
