+++ utils/config/db.ts
@@ -1,15 +1,15 @@
-// import mongoose from "mongoose";
+import mongoose from "mongoose";
 
-// const connectDB = async () => {
-//   try {
-//     await mongoose.connect(process.env.MONGO_URI!, {
-//       autoIndex: true
-//     });
-//     console.log("connected to mongodb");
-//   } catch (error) {
-//     console.error(`Error ${error}`);
-//     process.exit(1);
-//   }
-// };
+const connectDB = async () => {
+  try {
+    await mongoose.connect(process.env.MONGO_URI!, {
+      autoIndex: true,
+    });
+    console.log("connected to mongodb");
+  } catch (error) {
+    console.error(`Error ${error}`);
+    process.exit(1);
+  }
+};
 
-// export default connectDB;
+export default connectDB;
